requirements.txt
python-dotenv
selenium
allure-python-commons
pytest==8.4.2
selene==2.0.0rc9
allure-pytest==2.15.0
pydantic_settings